--- src/ome_zarr_models/_v06/axes.py
@@ -1,30 +1,0 @@
-from typing import Literal
-
-from pydantic import JsonValue
-
-from ome_zarr_models.base import BaseAttrs
-
-__all__ = ["Axis", "AxisType"]
-
-
-AxisType = Literal["space", "time", "channel"]
-
-
-class Axis(BaseAttrs):
-    """
-    Model for an element of `Multiscale.axes`.
-    """
-
-    # Explicitly name could be any JsonValue, but implicitly it must match Zarr array
-    # dimension_names which limits it to str | None
-
-    name: str | None
-    type: (
-        Literal["array", "space", "time", "channel", "coordinate", "displacement"]
-        | str
-        | None
-    ) = None
-    discrete: bool | None = None
-    # Unit probably intended to be str, but the spec doesn't explicitly specify
-    unit: str | JsonValue | None = None
-    longName: str | None = None
