+++ tests/_v06/test_image_label.py
@@ -1,7 +1,6 @@
 from zarr.abc.store import Store
 
-from ome_zarr_models._v06.axes import Axis
-from ome_zarr_models._v06.coordinate_transforms import CoordinateSystem, Scale
+from ome_zarr_models._v06.coordinate_transforms import Axis, CoordinateSystem, Scale
 from ome_zarr_models._v06.image_label import ImageLabel, ImageLabelAttrs
 from ome_zarr_models._v06.image_label_types import Color, Label, Source
 from ome_zarr_models._v06.multiscales import Dataset, Multiscale
