+++ tests/_v06/test_multiscales.py
@@ -1,9 +1,9 @@
 import pytest
 from pydantic import ValidationError
 
-from ome_zarr_models._v06.axes import Axis
 from ome_zarr_models._v06.coordinate_transforms import (
     AnyTransform,
+    Axis,
     CoordinateSystem,
     Scale,
     Sequence,
